Fix warning about excerpt separator
By default, the excerpt separator in Jekyll is a double newline. This is mainly for blogs that want to list the latest posts, but only show an excerpt.

To determine how much of the beginning of the post to show in the excerpt, you define the stopping point by adding a double newline after it. Or you can specify your own custom separator.

In a couple of our posts, namely `user-info.md` and `certificates.md`, the post content starts with a Liquid tag, and the content between the Liquid tags contains double newlines (i.e. an empty line between paragraphs). If excerpts were used, this would result in invalid Liquid syntax because the excerpt output would be missing the end Liquid tag.

To overcome this, Jekyll inserts the closing tag to appease Liquid, but the output might be missing content, hence the warning.

To get rid of this warning, I defined a custom excerpt separator in `_config.yml`

References:
https://jekyllrb.com/docs/posts/#post-excerpts

https://talk.jekyllrb.com/t/whats-the-deal-with-found-a-liquid-block-containing-the-excerpt-separator-n-n/5289/4