Update JWT docs for outgoing security events
According to the idp code, we are sending the `kid` header for outgoing security events, and we also include the `exp` claim. However, those were not documented, and we had a partner ask us why we don't send the `kid` header. They assumed that we don't because they believed the documentation.